Implement ECMA-402 related functions
* Fix some bugs related to Intl
* Use default timezone from VM instance
* The following functions should use Intl when the js engine supports Intl.
  Number.prototype.toLocaleString ([ locales [ , options ]])
  Date.prototype.toLocaleString ([ locales [ , options ]])
  Date.prototype.toLocaleDateString ([ locales [ , options ]])
  Date.prototype.toLocaleTimeString ([ locales [ , options ]])
* This patch fixes #251
